updated model with datetime, migrated and updated database. Started to test database by adding things to it in program.cs

+++ Models/Track.cs
@@ -4,7 +4,7 @@
     {
         public int TrackId { get; set; }
         public string Title { get; set; }
-        public int Length { get; set; }
+        public float Length { get; set; }
 
         public int VinylId { get; set; }
         public Vinyl Vinyl { get; set; }
